requirements.txt
argh==0.26.2
blessings==1.6
bpython==0.16
certifi==2017.4.17
chardet==3.0.4
click==6.7
curtsies==0.2.11
get==0.0.0
greenlet==0.4.12
idna==2.5
mock==2.0.0
pathtools==0.1.2
pbr==3.1.1
post==0.0.0
public==0.0.0
py==1.4.34
Pygments==2.2.0
pytest==3.1.3
python-decouple==3.0
PyYAML==3.12
query-string==0.0.0
request==0.0.0
requests==2.18.1
setupfiles==0.0.50
six==1.10.0
urllib3==1.21.1
watchdog==0.8.3
wcwidth==0.1.7
